secure-thrift: Fix memory and CPU usage in fallback mode
Summary:
D2567582 introduced more memory allocations and memcpy()s on the first
message on an encrypted channel in order to try falling back to insecure.

This diff limits these allocations and copies to the case where this first
message looks like a valid header message (i.e. contains the header magic).

In other words, allocations and copies will only happen on the first message on
an encrypted channel if it contains the header magic (because the client has
fallen back to insecure after a timeout, but the server thinks it's a secure
channel). In most cases this will be a real header message, in which case the
copy and memcpy is unavoidable. In rare cases, we'll also have the copies when
the encrypted message contains the header magic 0x0fff at the correct place by
chance. In these cases, everything will still be handled correctly, but with
this extra copy. I doubt (but haven't checked) that the encrypted krb message
format allows the bytes 0x0fff at the beginning.

Test Plan:
unittests
Reviewed By: haijunz@fb.com

Subscribers: alandau, yfeldblum

FB internal diff: D2584038

Signature: t1:2584038:1445971381:6252df3b0a1dcf443eed707510b37a4328866927

Blame Revision: D2567582