fix nodejs server compression bug
Summary: Node server was not calling THeaderTransport::transform() to apply transforms to the message. Result is that the client got a response that was marked as compressed in the header, but was not really compressed.

Test Plan:
Sent an fb303 getStatus to a locally running node server before this fix => client crashed with an exception
Sent an fb303 getStatus to a locally running node server after this fix => client got ALIVE

Reviewed By: haijunz@fb.com, alandau@fb.com

Subscribers: alandau, yfeldblum

FB internal diff: D2612830

Tasks: 8954701

Signature: t1:2612830:1446592144:ffda2f4ce454174864f29378f9cd3e1f1d6ff1aa