thrift-reflection: Fix static vs non-static function declaration vs definition
Summary:
The prototype for the `reflectionInitializer_xxx` functions was
without static, but the definition was with static unless it's a struct.

This diff synchronizes the static-ness of the prototype with the definition,
and also fixes it for typedefs.

Test Plan: fbmake

Reviewed By: yfeldblum@fb.com

Subscribers: alandau, yfeldblum

FB internal diff: D2645086

Signature: t1:2645086:1447292187:5c468048db8e4bebbb082629fbf78e0b285f7b72