part 3: add shape_arraykeys thrift option for Key Coercion
Summary:
We want the to-php thrift compilation to add a

class A {
const type TShape = shape (...);

to each PHP file (when certain options are enabled).

This is the third diff in the series.

This diff is an attempt to work around the Key Coercion madness in PHP.

Suppose we have a thrift structure:

struct A {
1: map<string, string> m;
}

And we store, in configerator, a map-key with value 123 as a string:

exportIfLast(new A(m = {'123': 'hello world'}))

The `__toShape()` method in `A` will call:

$this->m->toArray()

which will convert the string '123' to an integer. For example,

hphpd> var_dump((Map {'123' => 'hello world'})->toArray())
array(1) {
[123]=>
string(11) "hello world"
}

This value will then (in the ConfigeratorShape singleton) fail type
validation because `A` has:

class A {
const type TShape = shape(
'm' => array<string, string>,
//           ^^^^^^^
//           ^^^^^^^
);

To get around this problem, in this diff, I have added a thrift option 'shape_araykeys' that
will instead write the TShape as

class A {
const type TShape = shape(
'm' => array<arraykey, string>,
//           ^^^^^^
//           ^^^^^^
);

The idea is: If your Thrift struct uses "integer keys" then turn on
shape_arraykeys and add some extra type validation to PHP. If not,
then leave this option off and you'll have regular string typing.

Test Plan: look at the fixtures.

Reviewed By: ckwalsh@fb.com

Subscribers: alandau

FB internal diff: D2630315

Tasks: 8982701

Signature: t1:2630315:1447180551:313e1017c865ccfab3864880dd53313bf144d8bf