Fix bug in ensureInteger
Summary:
`ensureInteger` has a bug in the inner loop that makes it skip over bytes, potentially requiring significantly more bytes than necessary (leading to quadratic runtime behaviour).

Line in question:

```
while (idx++ < avail.second) {
  ...
}
```

The issue is that when `idx == avail.second`, `idx` is still incremented (effectively skipping that byte). If that byte was actually the last byte in the varint, this means we keep going (and if the next value was another varint, this potentially leads to going forever).

Note that this never actually causes an error, since Thrift structures always end with a 0 byte in the compact encoding.

I also moved one of the if statements into the while condition for readability. The code is equivalent.

Reviewed By: haijunz

Differential Revision: D2688441

fb-gh-sync-id: 5b44b1a60788d9cfa223a999571936764109135b