Generated-code examples for constants, cpp2-style
Summary:
[Thrift] Generated-code examples for constants, cpp2-style.

In order to fix some issues with the cpp2-style generated code for constants, we need to have an example to work with. We already have a good example for cpp1-style generated code for constants, which also needs some of the same sort of fixing, so we can just appropriate and repurpose that example for cpp2.

This diff only adds cpp2-style fixtures. All hypothetical fixing would happen in a separate diff.

Reviewed By: meyering

Differential Revision: D2699293

fb-gh-sync-id: ebaabfe5e774c0a337a5d47ec4802f1870790aa1