Close Cpp2Connection when using an existing connection
Summary:
Cpp2Connections "own" themselves via a circular shared_ptr.  The ConnectionManager should shut down the Cpp2Connection, which causes it to stop itself, when the server shuts down.  However, with already-existing channels, the connection isn't managed by a ConnectionManager, so the Cpp2Connection is leaked.

This diff changes:

* we call Acceptor::init() before useExistingChannel(), so that the connection is managed by a ConnectionManager
* we call drainAllConnections() on ThriftServer::duplexWorker_ when dtor'ing, because the ThriftServer doesn't own or have a reference to the socket under that worker, so it has to clean it up manually
* throw in a few DCHECK's to make sure we're not using cleanUp() when working with an existing channel

This does have the side effect of definitely not allowing a DuplexChannel to be reused for another server, but I'm not sure that works anyways.

Reviewed By: djwatson

Differential Revision: D2673465

fb-gh-sync-id: 15c50a315f6f03281ee0db965d77225f9563163b