Rewrite HTTPClientChannel to make non-http1x codecs work
Summary:
Previously HTTPClientChannel only used a codec directly. While this
worked for the HTTP1x codec, it does not work for HTTP2 as that depends on flow
control and other operations that weren't handled properly.

This rewrite migrates to using the Proxygen session system which handles
transactions separately and does things right with regards to flow control.

Reviewed By: andrewcox

Differential Revision: D2570573

fb-gh-sync-id: b02235a7a6ba413a0244bc220f73f37fee2dd16d