Introducing ProxygenThriftServer
Summary:
This is a Thrift2 Server implemented on the Proxygen networking stack.
It is intended as a near drop in replacement for ThriftServer, and can be used
to expose an HTTP2 interface for thrift requests. For compatibility we may want support for other protocols in the future, including HTTP1 and thrift header protocol, but I leave those as an exercise for future diffs.

Comparable Supported Features:
* Binary and Compact Protocol (via header or content sniffing)
* Thread count configuration, IO and worker
* Timeouts
** Connection idle
** Request Timeout
*** Server side (soft)
*** Client side (hard)
* Headers
** Client load header
* Debug Counters
** Active request count
** Load (Request/Queue)
** Dropped Connections
* TServerObserver Support

New Features:
* Header Compression (HTTP2 HPACK)
* Flow Control (Per Connection and Per Request)

Non-Backwards Compatible Changes:
* SASL encryption
* Oneway Thrift methods are not supported (void are still fine)

Planned Features, not included in this diff:
* Thrift Header protocol
* SSL sockets
* Additional configuration for threads (Stack size, setting thread factory, etc)
* Failure Injection
* Connection Load
* Message body compression (gzip)
* Generator support to work with HTTPMessage objects directly

Reviewed By: mmcduff

Differential Revision: D2570632

fb-gh-sync-id: 81a4d4d874a4bb453d16e7856b41cd3c2d904a47