Clean up configure.ac
Summary:
Open source mcrouter is planning to add fbthrift as a dependency, and we are
planning to do a bit of refactoring in the fbthrift build so that we can build
serialization/deserialization as a small library, and C++ generation as a
separate library.  For ease of maintenance, let's start by cleaning up the
current fbthrift configure.ac a bit.

Reviewed By: djwatson

Differential Revision: D2736665

fb-gh-sync-id: 7c08398f9bac62223e315d58694df8e40b29a9b6