Specialize std::hash and std::equal_to for cpp2 enums
Summary:
[Thrift] Specialize `std::hash` and `std::equal_to` for cpp2 `enum`s.

The motivation is to enable use of `enum`s as hash-set elements and hash-map keys. Users can certainly make their own specializations, but they will all be pure boilerplate, so it seems silly for users to have to do that. Let us simply enable this use-case everywhere.

`std::hash` has no meaning, so implementing it for an `enum` is not questionable, and `std::equal_to` over an `enum` is quite sensible. Implementing `std::equal_to` doesn't actually help with hash-set and hash-map, but it was right there, so why not.

Reviewed By: alandau

Differential Revision: D2738027

fb-gh-sync-id: e925d0305872dcbe1b361adb50f78a76e1f641fc