Delete a useless comment
Summary:
sendReply and sendErrorWrapped can only be called from eb thread and
the user-visible apis in the processor always wraps the calls to them with
runInEventBaseThread. The only case this flag can be touched from other thread
is when cancelRequest is called, but it is only called when the connection is
closed.

Reviewed By: alandau

Differential Revision: D2742476

fb-gh-sync-id: bcf81f93e20834399525c4a8b282ad0339891cc4