set HEADER_FLAG_SUPPORT_OUT_OF_ORDER on SASL messages in client
Summary:
We set HEADER_FLAG_SUPPORT_OUT_OF_ORDER on all non-SASL messages below
in setRequestHeaderOptions. Also set it in SASL messages, to be compatible with
servers which want all messages on the same socket to agree. (Servers which
support SASL don't mind if they're different, but servers which don't support
SASL may care, and we need to support them so that thrift_security=permitted
can successfully fallback with these older servers.)

Reviewed By: alandau, mshneer

Differential Revision: D2733729

fb-gh-sync-id: ea3f59d18234e488432089c7ffc3ba5c5fa40dbb