Fix `./configure --without-cpp`
Summary:
Building with

   autoreconf -ivf
  ./configure --without-cpp && make && sudo make install

should build the Thrift cpp/cpp2 compilers, but not the cpp/cpp2 libs.
Currently, `./configure --without-cpp` errors out on my Ubuntu 14.04 VM.
This diff fixes that issue and allows one to build the Thrift cpp/cpp2 compilers
without requiring unnecessary dependencies.

Motivation: open source mcrouter will want to build only the Thrift compiler
(and a small portion of the cpp/cpp2 libraries, but that work is still to be
done).

Reviewed By: pavlo-fb

Differential Revision: D2748135

fb-gh-sync-id: 6900ca321497fba80c469a283ef19109883ca2b0