Fix use after free in ProxygenThriftServer::ThriftRequestHandler
Summary:
ProxygenRequest is owned by the processor in a unique pointer. If the
processor destroys it, and the ThriftRequestHandler that created it is still
alive, it should clear the raw pointer to itself from the handler. We already
clear the handler pointer inside ProxygenRequest when the ThriftRequestHandler
is destroyed.

These should be safe from race conditions since both should execute on the evb
thread.

Reviewed By: haijunz

Differential Revision: D2746579

fb-gh-sync-id: 256fec4a7ed2cdcb93ffcafb03cd3bd0cb8aabd4