add a __jsonArrayToShape() method
Summary:
This diff adds

  __jsonArrayToShape(array<arraykey,mixed> $json_data): ?self::TShape;

The primary motivation is to introduce a highly optimized json
deserializer that allows hot code-paths to replace

```
TSimpleJSONSerializer::deserialize(
  $raw_config,
  $thrift_object,
);

$shape = $thrift_object->__toShape();
```

with

```
$json_data = fb_json_decode($raw_config);

$shapeB = $thrift_object::__jsonArrayToShape($json_data);
```

This diff also adds a new thrift option:

  shape_unsafe_json

that, when enabled, disables the type validation in `__jsonArrayToShape()`.
This is needed for hot code-paths.

This diff also fixes a bug in __fromShape() where the arraykey option was
not properly being handled.

Reviewed By: ckwalsh

Differential Revision: D2748012

fb-gh-sync-id: 5452678e77fe3a2b670e455c8f6516e994790460