Recover exception type and what() if future_foo() throws std::exception
Summary:
Before, we'd go through the exception_wrapper(exception_ptr) constructor, which
(for efficiency) doesn't save the type and what() (because that would require
rethrowing), so the TApplicationException received on the client is empty
and confusing.

Catch std::exception and save the type and what(). If you're throwing something
else, you're still SOL, but that's to be expected.

Reviewed By: yfeldblum

Differential Revision: D2774687

fb-gh-sync-id: af35bd312e337de3e55cfb778e5df7a087911c94