Induce compile failures when serializing cpp2 structs with cpp1 serializer
Summary:
[Thrift] Induce compile failures when serializing cpp2 structs with cpp1 serializer.

Right now we can get link failures, which can be painful to track down. This diff is to turn those into compile failures.

We assume that every `T` that needs to be serializable using the cpp1 serializer inherits `TStructType<T>`. All the cpp1-generated structs do, but custom structs intended to be serialized with the cpp1 serializer don't. Those custom structs will need to add that base class.

Reviewed By: haijunz

Differential Revision: D2767063

fb-gh-sync-id: f2a4f954c2cd2845e9a0da636c280f87db0eaf50