JSON Protocol support for cpp2
Summary:
[Thrift] JSON Protocol support for cpp2.

The cpp2 library has has simple-json-protocol support for a while, but is missing json-protocol support. We add that support, along with a quick test case to verify basic compatibility with the cpp1 version of json-protocol.

To avoid duplicating too much code, we move the bulk of the simple-json-protocol implementation into a new pair of base classes (reader and writer), and have both the simple-json-protocol and json-protocol implementations derive it. Note: this is not a set of interface classes; nothing should ever instantiate them directly; they are only to avoid unnecessary code duplication.

Reviewed By: haijunz

Differential Revision: D2798527

fb-gh-sync-id: 35318c9398a13b56b1a33641a24e289060aed3a0