Fix exception in ProtectionHandler.cpp: Error unwrapping message
Summary:
When sending a Task Timeout exception, the server creates a new timeoutHeader_,
which is prepended to the message. The default clientType is HEADER (without
SASL), so when updateClientType() is called, it adds a thrift_auth=0 header to
this message (although the message itself stays encrypted).
When the client receives this message, it sees the header and sets its client
type to match (i.e. drops SASL) and sends the next message in clear text.
The server, of course, tried to decrypt this plaintext message and fails with

```
ProtectionHandler.cpp:116] Exception in ProtectionHandler::read():
apache::thrift::TKerberosException: Error unwrapping message Invalid token was
supplied; Token header is malformed or corrupt
```

Reviewed By: haijunz

Differential Revision: D2813809

fb-gh-sync-id: ed5a958c0095d8cbbe1aca314c7cd118d45ff688