fix a rare memleak-on-shutdown bug in KerbersosSASLHandshakeClient
Summary:
There's a rare memory leak that triggers ASAN test failures in KerbersosSASLHandshakeClient. The leak is caused by scheduling the cleanupState() function on the saslThreadManager_ while it is shutting down. The thread manager doesn't get join()'ed, so it's possible for scheduled tasks to never run.

The solution is to have a dedicated ThreadManager that's only used to run cleanupState() tasks. This is wrapped in a folly::Singleton and is join()'ed properly when the singleton instance is destructed.

Reviewed By: shamdor-fb, siyengar

Differential Revision: D2825106

fb-gh-sync-id: 88f708800f5a466da88c587c581472393c68626d