Avoid races for a port
Summary: There is a kernel feature for binding to an unused port, so it's silly to iterate over all the ports manually. Also, the current code has a race condition: multiple threads can all think a port is free, and then they throw an exception when everyone isn't able to bind to it at the same time.

Reviewed By: yfeldblum

Differential Revision: D2835704

fb-gh-sync-id: 1009d1181b57d6cfc0947d7d72b550c3843c606d