Handle a load header that is set but empty
Summary: Before December 2015, The old code checked to see whether the header exists. The new code checks to see whether it exists and is non-empty. It's legal for the header to be set to the empty string, and so we need to account for that. With this change, an empty load header value will still be tracked in a `folly::Optional`.

Reviewed By: alandau

Differential Revision: D2836830

fb-gh-sync-id: b5f0eb5dbc02a6337f463e7069c187c89c046b3d