in java gencode, construct enum maps with reflection
Summary:
In the generated code for enums, instead of defining the VALUES_TO_NAMES map
and VALID_VALUES set as an explit list of the constants, populate it at runtime
with java reflection. This is necessary to handle enums with a large number
of values, so that the `<clinit>` size doesn't get too big to fit in bytecode.

Reviewed By: yfeldblum

Differential Revision: D2858506

fb-gh-sync-id: 78cd870743c3ccbcd9d543e0b9ff9674d58f266d