Release request buffer earlier
Summary:
The request buffer is owned by the request in its entire lifecycle,
even after deserializeRequest is called. The only use case for this buffer is
in sendErrorWrapped where it is used to read the method name and protocol
sequence id. This diff saves method name and protocol sequence id in
per-request Cpp2RequestContext so the request buffer can be released after
deserializeRequest.

The first thing the generated processor does is call readMessageBegin to read
the method name and find the handler method for it. This step is moved to
Cpp2Connection::requestReceived so that when the request is passed to
processor, this is already done. This ensures custom processors still work.

Reviewed By: alandau

Differential Revision: D2767489

fb-gh-sync-id: 2fb05664bbc589a64df60383e8667314c038b43f