Refactor thrift.util.asyncio to facilitate code sharing between asyncio and trollius
Summary:
In order to support asyncio-like features in Thrift/Python 2, we would like to reuse most of the logic defined in
thrift.util.asyncio, however the common code needs to be backwards compatible with Python 2.6+. This revision enables
that by moving all of the code that could be shared into the base processor class: `TProcessor`.

It probably makes more sense to look at D2848376 to understand why this would be very useful.

Reviewed By: ambv

Differential Revision: D2853064

fb-gh-sync-id: 1a71636038a01ca1f1efe44f25bd30c710a89452