Fix FrozenRange comparison for non-ASCII characters
Summary:
We saw issues in prod when calling find() on a frozen map<string, ...> where
for certain keys find() would return end(), but the key would actually be
found if find() was called on the map thawed from the frozen one. So the data
was definitely in the frozen blob, but couldn't be retrieved using find().

I was suspecting an issue with the binary search and digged a little deeper.
The keys in question could be arbitrary strings, in particular they could
contain UTF-8 characters. The comparison function for FrozenRange is defined
in terms of lexicographical_compare(). However, it turns out that this is not
equivalent to string comparison:

  string s1("anxiety");
  string s2("añonuevo");
  string s3("aot");

  cout << (s1 < s2) << endl;
  cout << (s2 < s3) << endl;
  cout << (s1 < s3) << endl;
  cout << "---" << endl;
  cout << lexicographical_compare(s1.begin(), s1.end(), s2.begin(), s2.end()) << endl;
  cout << lexicographical_compare(s2.begin(), s2.end(), s3.begin(), s3.end()) << endl;
  cout << lexicographical_compare(s1.begin(), s1.end(), s3.begin(), s3.end()) << endl;

yields:

  1
  0
  1
  ---
  0
  1
  1

Presumably that's because lexicographical_compare() treats char as signed whereas
string compare treats it as unsigned.

As a result, the order in which the data is frozen (based on string compare) is
different from the order expected by the binary search in FrozenRange.

This change implements FrozenRange's comparison operators in terms of folly::Range,
which has the same semantics for string-like iterators as string compare.

Reviewed By: ddrcoder

Differential Revision: D2886115

fb-gh-sync-id: 8edbac02c75bf09378e7bbcc0eadd4a78341fa89