Fixing union setters for indirectly recursive types
Summary:
when there's a cyclic dependency between types (say, A contains B, B
contains A), some union setters may fail. This is one scenario where this
happens:

- A is a forward declared struct
- B is a union which has A as a member
- the definition of B's setter for A is written to the header, even though there's no layout information for A
- A is defined after the aforemention definition

The fix that this diff proposes is to move the definition of the setters for a union to the .tcc file.
Once the .tcc file is included, all the layout information is already there.

Reviewed By: yfeldblum

Differential Revision: D2857732

fb-gh-sync-id: 7392e4438dc9c1384f9f354c89b359477041788f