Kill isunevenload
Summary:
split from D2876988
This kills isUnevenLoad. The flag was completely misnamed (server doesn't control evenness of load, it depends on connections), and only one client actually used it. Further, while it was a 10% regression in my benchmarks, that was doing 2 million QPS. At 500k it was only 5%, and I didn't find any thrift servers doing over 100k QPS.
In terms of optimizing performance, this isn't that easy to get right. We could extrapolate based on one thread, but we'd probably want to have a reference to the current worker. We can't really do a write optimized counter, as if the server is doing load balancing, getLoad is called every request. So I left this for another diff.

Reviewed By: federico, haijunz

Differential Revision: D2881595

fb-gh-sync-id: 675f6d6c852992a6dc025c8b2b294261e411e34f