Make sure the transport channels are properly closed on py-asyncio + thrift
Summary:
When using asyncio with thrift, we have been using a thrift transport layer represented by `WrappedTransport`.
The are a couple of issues with the way we're finalizing it now:

1/ `WrappedTransport` is not canceling pending tasks from `SenderTransport`, which it delegates to. This is addressed by
this diff.

2/ The thrift client does not hold a reference to the thrift transport object for long enough. A new transport object is
created every time a new thrift request is sent with asyncio. Therefore, the transport object may be garbage collected
before the thrift response is processed. asyncio will output an error to the logger saying a pending task was being
gargabe collected while waiting on a future result.

This will be addressed in a separate diff as I think it will be easy to fix this once D2853198 lands because we could
simply instantiate one thrift transport object per thrift client, that would probably happend in
`AsyncClientProtocolBase::connection_made`, then only the thrift protocol object could be passed in as an argument to
our thrift client (just like the vanilla py thrift client, btw).

Reviewed By: ambv

Differential Revision: D2905863

fb-gh-sync-id: 2568c1d768c43f2d181fbef55021fd71529a57ed