Fix fallback to plaintext
Summary:
Fallback to plaintext happens in handleSecureMessage, when the client
thinks the sasl handshake fails but the server thinks it succeeded. So the
server will receive a plaintext request with client type set to header and
protection state is still VALID. So simply skipping handleSecureMessage when
protection state is VALID is a bug.

Reviewed By: alandau

Differential Revision: D2907010

fb-gh-sync-id: e43907141d2b209d7225dbdfaade6a6e7002b405