Passing an event loop from above to asyncio-based Thrift
Summary:
Let's open up the possibility of running thrift clients and processors in a custom event loop provided by the caller.

The current asyncio thrift code doesn't work if the caller is expecting it to run in a custom event loop, other than the
default one in the main thread (`asyncio.get_event_loop`).

Reviewed By: ambv

Differential Revision: D2916637

fb-gh-sync-id: 51dcd66f4286a261b8c9d5397a3fbcdccda6ed01
shipit-source-id: 51dcd66f4286a261b8c9d5397a3fbcdccda6ed01