add hs_include syntax
Summary:
add syntax to support included haskell files

In haskell, we have type classes. If you want a generated thrift datatype to be an instance of some type class, you need to implement that instance externally right now. This is called an orphan instance and it greatly increases the complexity of compiling the code. When you turn on orphan instances, the compiler has to look all over the place in order to resolve type class constraints as opposed to only looking in a single file.

We therefore need a way of injecting additional code into the generated module

Of course, the injected code could have type errors. Luckily, haskell has these nice {-# LINE #-} pragmas which move the locations of reported type errors. For example, if there is an error in some code that got spliced into a generated module, the type error will point you to the original source file where you wrote the code. This makes debugging a lot easier.

Reviewed By: yfeldblum

Differential Revision: D2921676

fb-gh-sync-id: 997b6216c8acf79ae0cbb628253fa680bcb59a69
shipit-source-id: 997b6216c8acf79ae0cbb628253fa680bcb59a69