A factory for py.asyncio thrift clients
Summary:There is coupling between `ThriftHeaderClientProtocol` and how thrift clients are created in py.asyncio.

See method `ThriftHeaderClientProtocol::connection_made` and the signature for the thrift client constructor. Note how
we're passing two factories as parameters to the thrift client object (to be created) so that on **each** thrift method
call, **new** thrift transport and protocol objects will be created.

First, afaik we could create single transport/protocol objects per thrift client, so that we avoid paying that overhead
on each thrift method call. Second, that coupling is not desirable imho because it ties `ThriftHeaderClientProtocol`
with a particular choice of thrift transport and protocol types.

Therefore, I propose we move the responsibility of creating thrift transport and protocol objects out of
`ThriftHeaderClientProtocol`. That's how `AsyncThriftFactory` came up, along with its specialization classes:
`AsyncioThriftFactory` and `TrolliusThriftFactory`.

Besides, this approach enables us to simplify the thrift client constructor given that it requires only two parameters
now: a thrift protocol instance and an asyncio event loop.

Reviewed By: fried

Differential Revision: D2923065

fb-gh-sync-id: 03f0457b40a7eca7a74d520bd24bf93721acbcda
shipit-source-id: 03f0457b40a7eca7a74d520bd24bf93721acbcda