Dyno thrift logging - measure CPU time from all ThreadManager threads
Summary: Basic idea: we want to measure total cpu time per request, from all threads - we add RequestPerfData with an atomic counter to RequestContext and add hooks to ThreadManager::Observer to run before and after a task is run on a thread. In the Observer pre-run hook, we check RequestContext and if there's RequestPerfData, we start a timer, stop it when the task finishes and update RequestPerfData.

Reviewed By: mmcduff, haijunz

Differential Revision: D2918295

fb-gh-sync-id: 00a15c5a0440df8e3811907c68e938882a6186dc
shipit-source-id: 00a15c5a0440df8e3811907c68e938882a6186dc