Fix Hashable instances for unordered containers
Summary:The toList function of HashMap or HashSet doens't guarantee order.
Actually, when there are hash collisions, the order of these elements
depends on the order they are inserted. As a result, a simple foldl
will result in different hash value for things that are equal (==).

The recent unordered-containers has implemented more efficient
Hashable instances for these types. But as we don't have access
to internal data type, let's fix this bug by implementing a less
efficient hash function.

Reviewed By: kuk0

Differential Revision: D3030816

fb-gh-sync-id: 93609e905298068a524f6077182f2b1cad0a0966
shipit-source-id: 93609e905298068a524f6077182f2b1cad0a0966