Ignore services in a principal with no realm.
Summary:Something is inserting "corrupted" entries in the credential cache files for the entries in hg.vvv and svn.vip.vvv, of the form:

```
Ticket cache: FILE:/tmp/ticket_mortimer_broken
Default principal: mortimer@DEV.FACEBOOK.COM

Valid starting       Expires              Service principal
03/07/2016 14:37:09  03/08/2016 00:37:09 krbtgt/DEV.FACEBOOK.COM@DEV.FACEBOOK.COM
03/07/2016 14:45:10  03/08/2016 00:37:09 host/hg.vvv.facebook.com@
03/07/2016 14:45:10  03/08/2016 00:37:09 host/hg.vvv.facebook.com@DEV.FACEBOOK.COM
```

so when the fbcode Kerberos libraries read the credential caches, they try to get a krbtgt for that weird realm ('') and the server returns UNKNOWN_SERVER, so the user gets:

```
E0724 10:49:08.000000 826391 thrift/lib/cpp/util/kerberos/Krb5CredentialsCacheManager.cpp:168] Failure in
credential cache thread: Server krbtgt/PRN1.PROD.FACEBOOK.COM not found in Kerberos database while get
credentials. If the application is authenticating as a user, run "kinit" to get new kerberos tickets. If
the application is authenticating as a service identity, make sure the keytab is in the right place and is
accessible. Check for bad format in kerberos config too.
```

which is leading to incredible churn in the support groups, and a big confusion among engineers.

The patch just discards any ticket in a cache with an empty realm.

NOTE: After this lands, we will need to rebuild any affected application to get a new version of the Kerberos libraries embedded in its binaries.

Reviewed By: abajenov

Differential Revision: D3039003

fb-gh-sync-id: cff5621cb3dfcd9eb7ed847cde46d03b4d41262c
shipit-source-id: cff5621cb3dfcd9eb7ed847cde46d03b4d41262c