Reverted commit D3014457
Summary:This diff adds a pcap logging capability to ThriftServer.
When enabled, ThriftServer will log (some) reuqests and responses to a file in
pcap format (same format used by tcpdump and wireshark).

This is not a network capture (as done by tcpdump), but an application-level
capture that mimics network packets.

Features:
 - Rotates pcap files after certain amount of data stored
 - Limit capture to first N bytes (to limit file size)
 - Only log first N messages at the start of a connection and only M last
   messages before it's closed
 - Stores plaintext messages (after decryption or before encryption)
 - pcap files created with 0600 permissions
 - Kill switch - Gflag that prohibits enabling of logging
 - Fake MAC addresses for packets: Last byte will be 00 for plaintext
   messages, 01 for Kerberos-encrypted messages, 02 for TLS-encrypted messages.
   Other bytes identify client or server (in the TCP sense) - client will have
   then set to 11, server to 22.
 - Only log a sample percent of connection

Design:
PcapLoggingHandler sits in the Cpp2Channel pipeline, between the Krb decryption
and Thrift unframing stages. Upon receiving a message (just received or ready
to be transmitted), it checks if logging is enabled. If not, just forwards to
the next pipeline stage. If yes, forwards through a MPMCQueue to a worker
thread that will process the message further. Writing to MPMCQueue is
nonblocking, if the queue is full, the message is dropped to avoid stalling the
IO thread.

The worker thread reads messages from this MPMCQueue, dumps the first N
messages on a connection to a file, stores the other in a circular buffer (of
size M) and writes them to the file on connection close.

Memory usage:
The first N messages are written to the file immediately and don't incur any
additional memory overhead. The last M messages have to be stored until
connection close (or we won't know that they are the last). Note that we only
store the last M messages, not all messages. Moreover, we only store the first
snaplen bytes from a message (by manipulating the IOBuf, and without copying
the data).
Due to the way IOBufs work, the whole IOBuf (not IOBuf chain) will remain in
memory, and not just the first snaplen bytes. However, currently, on receive,
big messages are split to 2 IOBufs in the chain (of size 2K and the rest) so at
most 2K will be stored until connection close. On the transmit path, the chain
is composed of 16K IOBufs, so at most 16K per message (for up to M messages)
will be stored until connection close.

On a loadgen test with QPS of 20k, consisting of small encrypted messages, the
memory overhead for storing the last 8 messages on a connection was 2MB (RSS
increased from 39MB to 41MB, as reported by top).

Reviewed By: haijunz

Differential Revision: D3014457

fb-gh-sync-id: 3952abd8b827c4edaf01b677a07cecc022e06203
shipit-source-id: 3952abd8b827c4edaf01b677a07cecc022e06203