properly setting `__isset` on thrift constants
Summary:is_set information for members of constants defined in a thrift file were unreliable.
this diff properly sets the is_set information for those fields, setting the flag for fields with explicitly defined values, and leaving if off for the rest.
A desirable side-effect of this diff is that Thrift no longer uses the fragile constructor anymore - now there's a selective constructor that allows one to explicitly specify which fields to set during construction - therefore it is ready for deprecation.

Reviewed By: alandau

Differential Revision: D2996302

fb-gh-sync-id: ba22c9d38ae8d7ae1e1e6c4c0953d14f956f6628
fbshipit-source-id: ba22c9d38ae8d7ae1e1e6c4c0953d14f956f6628