adding `reflect_struct::member`
Summary:just like `reflect_struct::names`, this is a shortcut to fetch the corresponding `reflected_struct_data_member` when the member name is known at compile-time. This avoids having to perform map lookups and makes it much more convenient to refer to members' metadata.
One planned use of this is to replace fragile constructors:

  using info = reflect_struct<MyStruct>;
  MyStruct instance(wrap_argument<info::member<>::some_int>(10), wrap_argument<info::member<>::some_string>("test"));

or even a friendlier version

  using arg = selective_constructor_argument<MyStruct>;
  MyStruct instance(arg::some_int(10), arg::some_string("test"));

Reviewed By: yfeldblum

Differential Revision: D3112145

fb-gh-sync-id: 498f496d410b00b362311d27508c341af7406599
fbshipit-source-id: 498f496d410b00b362311d27508c341af7406599